Fix distributed logger bug
